mergeNode2Children method created
Method created for Node2, Node3, and Node4. For Node2, the node itself
is consumed in the merge, but for Node3 and Node4, a key is taken as a
parameter and consumed during the merge with the two Node2 children.
Please call these correctly I think it will destroy everything if we
don't hahahha